requirements.txt
django
whitenoise
psycopg2-binary
djangorestframework
django-polymorphic
jsonfield